requirements.txt
Flask==3.0.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.2
kaleido>=0.2.1
gunicorn>=20.1.0
